third_party/WebKit/LayoutTests/fast/dom/shadow/distribution-update-recalcs-style.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<script>
function getColorProperty(id) {
    return window.getComputedStyle(document.getElementById(id)).color;
}

function log(message) {
    var console = document.getElementById('console');
    console.innerHTML += message + '\n';
}
</script>
</head>
<body>
<pre id="console"></div>
<div id="test">
    <div id="host" style="color: red">
        <div id="div-child"></div>
        <span id="span-child"></span>
    </div>
</div>
<script>
description("Tests to ensure that updating the select rule in projection causes style recalc.");

var sr = document.getElementById("host").createShadowRoot();
sr.innerHTML = '<style>div { color: blue }</style><div><content id="contentId" select="#div-child"></content></div>';

shouldBe('getColorProperty("div-child")', '"rgb(0, 0, 255)"');
shouldBe('getColorProperty("span-child")', '"rgb(0, 0, 0)"');

sr.getElementById("contentId").select = "#span-child";
shouldBe('getColorProperty("div-child")', '"rgb(0, 0, 0)"');
shouldBe('getColorProperty("span-child")', '"rgb(0, 0, 255)"');

if (window.testRunner)
    testRunner.notifyDone();
</script>
</body>
</html>
