third_party/WebKit/LayoutTests/fast/dom/shadow/distribution-update-recalcs-style-expected.txt

Tests to ensure that updating the select rule in projection causes style recalc.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".




    


PASS getColorProperty("div-child") is "rgb(0, 0, 255)"
PASS getColorProperty("span-child") is "rgb(0, 0, 0)"
PASS getColorProperty("div-child") is "rgb(0, 0, 0)"
PASS getColorProperty("span-child") is "rgb(0, 0, 255)"



PASS successfullyParsed is true

TEST COMPLETE
